lib/docker/requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
awscli==1.33.22
blinker==1.6.2
boto3==1.34.140
botocore==1.34.140
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
dataclasses-json==0.6.1
docutils==0.16
exceptiongroup==1.1.3
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.41
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.1.11
langsmith==0.1.17
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
opensearch-py==2.3.1
packaging==23.2
pandas==2.1.1
Pillow==10.2.0
pip==23.3
protobuf==4.24.3
pyarrow==14.0.1
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pypdf==3.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
rsa==4.7.2
s3transfer==0.10.2
setuptools==65.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
streamlit==1.35.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.18
validators==0.22.0
wheel==0.38.4
yarl==1.9.2
zipp==3.17.0
